requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
arxiv==2.2.0
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
composio_core==0.7.15
composio_openai==0.7.15
cryptography==44.0.3
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
litellm==1.67.5
magika==0.6.2
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.8.1
mcpo==0.0.14
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.2.5
onnxruntime==1.22.0
openai==1.78.1
openai-agents==0.0.14
packaging==25.0
paramiko==3.5.1
parso==0.8.4
passlib==1.7.4
pdfminer.six==20250506
pexpect==4.9.0
pip==23.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pymupdf4llm==0.0.24
PyNaCl==1.5.0
pyparsing==3.2.3
pyperclip==1.9.0
Pysher==1.0.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
pyzotero==1.6.11
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
semver==3.0.4
sentry-sdk==2.28.0
setuptools==65.5.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.3
types-requests==2.32.0.20250328
typing_extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uv==0.7.5
uvicorn==0.34.2
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.20.0
zipp==3.21.0
zotero-mcp==0.1.6
aiosqlite==0.19.0
python-multipart==0.0.20
